backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
python-magic==0.4.27

# Database
sqlalchemy==2.0.35
alembic==1.13.3

# Audio processing
openai-whisper==20231117
pyannote.audio==3.3.2
ffmpeg-python==0.2.0
pydub==0.25.1

# Background tasks
celery==5.4.0
redis==5.2.0

# LLM integrations
httpx==0.27.2
openai==1.54.3

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==5.0.0
httpx==0.27.2

# Code quality
black==24.10.0
ruff==0.7.3
mypy==1.13.0

# Utilities
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.6.0
